Revert "Tests/add rspec"
