+++ Gemfile
@@ -3,9 +3,3 @@ source "https://rubygems.org/"
 gem 'mechanize'
 gem 'nokogiri'
 gem 'rest-client'
-
-group :test do
-  gem 'webmock'
-  gem 'rspec'
-  gem 'vcr'
-end
